pyproject.toml
[build-system]
requires = ['poetry-core>=1.2.0', 'wheel',]
build-backend = 'poetry.core.masonry.api'


[tool.poetry]
name = 'complyscribe'
version = "0.13.0"
description = "complyscribe assists users in leveraging Compliance-Trestle in automated workflows of for OSCAL formatted compliance content management."

authors = ["Jennifer Power <jpower@redhat.com>",]

include = ['LICENSE']
exclude = ['tests/', 'docs/']
license = 'Apache-2.0'
readme = 'README.md'

repository = 'https://github.com/RedHatProductSecurity/complyscribe'


[tool.poetry.scripts]
complyscribe = "complyscribe.cli.root:root_cmd"

[tool.poetry.dependencies]
python = '^3.9.2'
gitpython = "^3.1.45"
compliance-trestle = "^3.9.1"
github3-py = "^4.0.1"
python-gitlab = "^6.2.0"
ruamel-yaml = "^0.18.14"
pydantic = "^2.11.7"
ssg = {git = "https://github.com/ComplianceasCode/content"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
black = "^25.1.0"
mypy = "^1.17.1"
isort = "^5.13.2"
safety = "^3.2.8"
flake8-print = "^5.0.0"
pre-commit = "^4.3.0"
mkdocs-material = "^9.6.16"
types-pyyaml = "^6.0.12.20240917"
mkdocs-include-markdown-plugin = "^7.1.6"

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.2.1"
pytest-skip-slow = "^0.0.5"
responses = "^0.25.8"

[tool.poetry.group.plugins]
optional = true

[tool.poetry.group.plugins.dependencies]
compliance-trestle-fedramp = "^0.4.0"

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    'tests/*',
]

[tool.pytest.ini_options]
minversion = '6.0'
addopts = """
--doctest-modules \
--cov=./ \
--cov-append \
--cov-report html:tests/reports/coverage-html \
--cov-report xml:tests/reports/coverage.xml \
--ignore=docs/
"""
testpaths = [
    'tests',
]
markers = [
    "integration: mark a test as an integration test.",
]


[tool.mypy]

plugins = [
  "pydantic.mypy"
]

follow_imports = "skip"
warn_redundant_casts = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = "github3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ruamel"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "responses"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ssg.*"
ignore_missing_imports = true
